Updates: scopes, dark mode, choice args, start of reference doc
* Files can now have a scope.

  !scope cloak

Sets the scope to 'cloak' at the top of a file. This means node names
like =dropcloak= in a file are read as =cloak:dropcloak=. 'g' is the
global scope, and games start at '=g:0=' by default. The fully
qualified form is used to reference a node in another file.

* Add support for dark mode when @media (prefers-color-scheme: dark)

* Add support for choice args. Example:

  =here=

  | const delta = 1; // pretend this was computed

  >answer the answer is 42; 42
  >answer the anser is at least 21; `${21 + delta}`

  =answer=

  The answer is ${choice.data}, but did we ask the
  right question?

* Make some minor changes to Cloak of Darkness to
  bring it closer to spec. https://mipmip.org/IFrescue/rf/

* Add some font-lock face improvements for 9l-mode.el

GitOrigin-RevId: 23a8367c29a23992a5b882806194cc907fe4c888